COMP: Getting rid of conversion warnings on 64-bit machines.

+++ VTK/Infovis/vtkDelimitedTextReader.cxx
@@ -376,7 +376,7 @@ splitString(const vtkStdString& input,
     }
 
   results.push_back(currentField);
-  return results.size();
+  return static_cast<int>(results.size());
 }
 
 // ----------------------------------------------------------------------
